pyproject.toml
[project]
name = "formallymad"
version = "0.1.0"
description = "Autonomous coding using the Formally MAD framework."
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "qbaf-py @ git+https://github.com/TimKam/Quantitative-Bipolar-Argumentation.git",
    "llmshap[all]",
]

[project.optional-dependencies]
visualize = [
    "qbaf-py @ git+https://github.com/TimKam/Quantitative-Bipolar-Argumentation.git#egg=qbaf-py[Visualizer]",
    "matplotlib",
]

test = [
    "pytest>=8.0.0",
]


[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
